refactoring auth logic, minor fixes in backend & extending contacts schema

+++ backend/src/controllers/users-controllers.ts
@@ -1,110 +1,16 @@
-import { Request, Response, NextFunction } from 'express';
+import { Request, Response, NextFunction } from "express";
 
-import User from '../models/User';
-import ExpressError from '../utils/ExpressError';
-import verifyRefreshToken from '../utils/verifyRrefreshToken';
-
-import { UserType } from '../Types/user-types';
-import { UserModel } from '../models/User';
-
-import generateUserAuthData from '../utils/generateUserAuthData';
-import mongoose from 'mongoose';
+import User from "../models/User";
+import ExpressError from "../utils/ExpressError";
 
 const user = new User();
 
-export const login = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
-	try {
-		const { password, email } = req.body;
-
-		const user = (await UserModel.findAndVerifyPassword(
-			email,
-			password
-		)) as UserType;
-
-		const authData = generateUserAuthData(user, true);
-
-		await UserModel.findAndUpdateRefreshToken(
-			user._id,
-			authData?.refreshToken as string,
-			{ isAdd: true }
-		);
-
-		res.status(200).json({
-			success: true,
-			data: authData,
-		});
-	} catch (err: unknown) {
-		if (err && err instanceof (ExpressError || Error)) {
-			next(err);
-		}
-	}
-};
-
-export const refreshToken = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
-	try {
-		const { refreshToken } = req.body;
-		const { id } = req.params;
-
-		const payload = verifyRefreshToken(refreshToken);
-
-		console.log('paylaod controller', payload);
-
-		const authData = generateUserAuthData(
-			payload as unknown as UserType,
-			false
-		);
-
-		await UserModel.findAndUpdateRefreshToken(
-			id as unknown as mongoose.Types.ObjectId,
-			refreshToken,
-			{
-				isReplace: true,
-				newToken: authData?.refreshToken as string,
-			}
-		);
-
-		res.status(202).json({ success: true, data: authData });
-	} catch (err: unknown) {
-		if (err && err instanceof Error) {
-			next(err);
-		}
-	}
-};
-
-export const logout = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
-	try {
-		const { id } = req.params;
-		await UserModel.findAndUpdateRefreshToken(
-			id as unknown as mongoose.Types.ObjectId,
-			undefined,
-			{ isEmpty: true }
-		);
-		res.status(200).json({ success: true, message: 'logged out successfully' });
-	} catch (err: unknown) {
-		if (err && err instanceof Error) next(err);
-	}
-};
-
-export const newUser = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
+export const newUser = async (req: Request, res: Response, next: NextFunction) => {
 	try {
 		const createdUser = await user.createUser(req.body);
-		res.status(202).json({ success: true, data: createdUser });
+		if (createdUser) return res.status(202).json({ success: true, data: createdUser });
+
+		throw new ExpressError("something went wrong, user is not created", 400);
 	} catch (err: unknown) {
 		if (err && err instanceof Error) {
 			next(err);
@@ -112,31 +18,24 @@ export const newUser = async (
 	}
 };
 
-export const findUser = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
+export const findUser = async (req: Request, res: Response, next: NextFunction) => {
 	try {
-		const { id } = req.params;
-		const foundUser = await user.getUser(id);
-		res.status(202).json({ success: true, data: foundUser });
+		res.status(202).json({ success: true, data: res.locals.user });
 	} catch (err: unknown) {
 		if (err && err instanceof Error) {
 			next(err);
 		}
 	}
 };
-export const deleteUser = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
+export const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
 	try {
 		const { id } = req.params;
+		if ((res.locals.user && res.locals.user._id === id) || res.locals.user.admin) {
+			const deletedUser = await user.deleteUser(id);
+			return res.status(202).json({ success: true, data: deletedUser });
+		}
 
-		const deletedUser = await user.deleteUser(id);
-		res.status(202).json({ success: true, data: deletedUser });
+		throw new ExpressError("not authorized to delete this account!", 403);
 	} catch (err: unknown) {
 		if (err && err instanceof Error) {
 			next(err);
